+++ model/artist.go
@@ -12,7 +12,7 @@ type Artist struct {
 	OrderArtistName string `json:"orderArtistName"`
 
 	// Annotations
-	PlayCount int       `json:"playCount"   orm:"-"`
+	PlayCount int64     `json:"playCount"   orm:"-"`
 	PlayDate  time.Time `json:"playDate"    orm:"-"`
 	Rating    int       `json:"rating"      orm:"-"`
 	Starred   bool      `json:"starred"     orm:"-"`
